requirements.txt
Django>=5.0,<6.0
djangorestframework>=3.14.0
djangorestframework-simplejwt>=5.3.0
django-cors-headers>=4.3.0
drf-yasg>=1.21.7
psycopg2-binary>=2.9.9
pandas>=2.0.0
python-dotenv>=1.0.0
dj-database-url
whitenoise
celery>=5.3.0
redis>=4.5.0
django-redis>=5.4.0
channels>=4.0.0
channels-redis>=4.2.0
daphne>=4.0.0